Rename schemas/call_participants -> call_participant
