+++ imports/client/components/CallLinkBox.tsx
@@ -13,7 +13,7 @@ import Flags from '../../flags';
 import { getAvatarCdnUrl } from '../../lib/discord';
 import CallSignals from '../../lib/models/call_signals';
 import Profiles from '../../lib/models/profiles';
-import { CallParticipantType } from '../../lib/schemas/call_participants';
+import { CallParticipantType } from '../../lib/schemas/call_participant';
 import { CallSignalType, CallSignalMessageType } from '../../lib/schemas/call_signals';
 import { ProfileType } from '../../lib/schemas/profiles';
 import { RTCConfigType } from '../rtc_config';
