+++ imports/client/components/RTCDebugPage.tsx
@@ -11,7 +11,7 @@ import CallParticipants from '../../lib/models/call_participants';
 import CallSignals from '../../lib/models/call_signals';
 import Profiles from '../../lib/models/profiles';
 import Puzzles from '../../lib/models/puzzles';
-import { CallParticipantType } from '../../lib/schemas/call_participants';
+import { CallParticipantType } from '../../lib/schemas/call_participant';
 import { CallSignalType } from '../../lib/schemas/call_signals';
 import { ProfileType } from '../../lib/schemas/profiles';
 import { PuzzleType } from '../../lib/schemas/puzzles';
